Start to revise pre-processing
* Put all required scripts in one workflow
* Include data import in pre-procesing
* Use a .yml file to specify scripts
* Allow bash scripts
* First examplary script: OSM import
