--- dataprocessing/tools/osm2pgsql/osm2oedb.txt
@@ -1,1 +1,0 @@
-osm2pgsql --create --slim --hstore-all --number-processes 2 --cache 5120 -H localhost -P 5432 -d oedb -U  -W  -p osm_deu -S oedb.style  germany-161001.osm.pbf
