+++ preprocessing/openstreetmap/default.style
