+++ preprocessing/openstreetmap/empty.style
